pyproject.toml
[tool.poetry]
name = "opencoursebot"
version = "0.1.0"
description = "A handy discord bot for college-related functionality such as information about classes, professors, and general stats."
authors = ["David Tso <davidtso1219@gmail.com>", "Madhav Varshney <madhavarshney@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.6.0"
beautifulsoup4 = "^4.9.3"
flask = "^1.1.2"
python-dotenv = "^0.15.0"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
